don't validate spare bits content in received GTPv1 messages
3GPP TS 29.060 is inconsistent with the handling of spare bits in
message elements. For some it specifies that they "...  shall be
set to 0 by the sending side and which shall not be evaluated by the
receiving side.", for others it defines the spares to be 1 (one).

It seems that some existing implementation have applied the first
version to all spare bits, even the once defined to be 1.

Change our logic to:
 * ignore the value of all spare bits regardless of the definition
 * set the ones defined as 1 to that value, the others to 0